fastbootd: Allow using kernel commits to pass SafetyNet
When using kernel commits to pass SafetyNet and CTS like:

  proc: Remove verifiedbootstate flag from /proc/cmdline
  proc: Set androidboot.verifiedbootstate=green

fastbootd returns the device as locked when flashing images
or resizing partitions when flashing because it can't find
androidboot.verifiedbootstate=orange in the kernel command
line letting it know it's safe to flash.

So if androidboot.verifiedbootstate is missing from the kernel
command line or if is returning verifiedbootstate=green report
the device as unlocked.

PS2: Add support for androidboot.verifiedbootstate=green

Change-Id: I56712f374ea26397ea5a183cfe100cf92e21d531
Signed-off-by: Jason Edson <jaysonedson@gmail.com>
Signed-off-by: Joey Huab <joey@evolution-x.org>

Former-commit-id: b6cdaea2192596c979a58bd95b24b65a424ea3c8
Signed-off-by: Joey Huab <joey@evolution-x.org>

Former-commit-id: 76dedf351a29b26b3476dd3d65fd891155daf4e8